requirements.txt
Django==1.6.5
South==0.8.4
django-localflavor==1.0
psycopg2==2.5.3
requests==2.3.0
wsgiref==0.1.2
